# swagger-editor/routes/index.js
const express = require('express');
const router = express.Router();

// Sample Hello World route
router.get('/', (req, res) => {
  res.send('Hello from the API!');
});

/**
 * @swagger
 * /projects:
 *   get:
 *     summary: Retrieve a list of all sustainability projects.
 *     parameters:
 *       - name: limit
 *         in: query
 *         description: Number of projects to retrieve.
 *         schema:
 *           type: integer
 *       - name: offset
 *         in: query
 *         description: Offset for pagination.
 *         schema:
 *           type: integer
 *     responses:
 *       200:
 *         description: A list of projects.
 *       400:
 *         description: Invalid query parameters.
 */
router.get('/projects', (req, res) => {
  // Implementation here
});

/**
 * @swagger
 * /projects/{id}:
 *   get:
 *     summary: Retrieve a specific project by ID.
 *     parameters:
 *       - name: id
 *         in: path
 *         required: true
 *         description: ID of the project to retrieve.
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Project details.
 *       404:
 *         description: Project not found.
 */
router.get('/projects/:id', (req, res) => {
  // Implementation here
});

/**
 * @swagger
 * /projects:
 *   post:
 *     summary: Create a new project.
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               name:
 *                 type: string
 *               description:
 *                 type: string
 *               category:
 *                 type: string
 *     responses:
 *       201:
 *         description: Project created.
 *       400:
 *         description: Invalid request body.
 */
router.post('/projects', (req, res) => {
  // Implementation here
});

/**
 * @swagger
 * /projects/{id}:
 *   put:
 *     summary: Update a project by ID.
 *     parameters:
 *       - name: id
 *         in: path
 *         required: true
 *         description: ID of the project to update.
 *         schema:
 *           type: string
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               name:
 *                 type: string
 *               description:
 *                 type: string
 *               category:
 *                 type: string
 *     responses:
 *       200:
 *         description: Project updated.
 *       400:
 *         description: Invalid request body.
 *       404:
 *         description: Project not found.
 */
router.put('/projects/:id', (req, res) => {
  // Implementation here
});

/**
 * @swagger
 * /projects/{id}:
 *   delete:
 *     summary: Delete a project by ID.
 *     parameters:
 *       - name: id
 *         in: path
 *         required: true
 *         description: ID of the project to delete.
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Project deleted.
 *       404:
 *         description: Project not found.
 */
router.delete('/projects/:id', (req, res) => {
  // Implementation here
});

/**
 * @swagger
 * /categories:
 *   get:
 *     summary: Retrieve all categories.
 *     responses:
 *       200:
 *         description: A list of categories.
 */
router.get('/categories', (req, res) => {
  // Implementation here
});

/**
 * @swagger
 * /categories/{id}:
 *   get:
 *     summary: Retrieve a specific category by ID.
 *     parameters:
 *       - name: id
 *         in: path
 *         required: true
 *         description: ID of the category to retrieve.
 *         schema:
 *           type: string
 *     responses:
 *       200:
 *         description: Category details.
 *       404:
 *         description: Category not found.
 */
router.get('/categories/:id', (req, res) => {
  // Implementation here
});

/**
 * @swagger
 * /ecobot:
 *   post:
 *     summary: Interact with the EcoBot chatbot.
 *     description: Send a message to the EcoBot chatbot and receive a response.
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               message:
 *                 type: string
 *                 description: The message to send to the chatbot.
 *     responses:
 *       200:
 *         description: Chatbot response
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 response:
 *                   type: string
 *                   description: The response from the chatbot.
 *       400:
 *         description: Invalid request body
 *       500:
 *         description: Server error
 */
router.post('/ecobot', (req, res) => {
  const { message } = req.body;
  // Chatbot logic here
  res.json({ response: `EcoBot received your message: ${message}` });
});

/**
 * @swagger
 * /translate:
 *   post:
 *     summary: Translate text to a specified language.
 *     description: Translate a given text into the desired language.
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               text:
 *                 type: string
 *                 description: The text to be translated.
 *               targetLanguage:
 *                 type: string
 *                 description: The language code to translate the text into.
 *     responses:
 *       200:
 *         description: Translated text
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 translatedText:
 *                   type: string
 *                   description: The translated text.
 *       400:
 *         description: Invalid request body
 *       500:
 *         description: Server error
 */
router.post('/translate', (req, res) => {
  const { text, targetLanguage } = req.body;
  // Translation logic here
  res.json({ translatedText: `Translated '${text}' to ${targetLanguage}` });
});

module.exports = router;